+++ files/ensembl/get_ensemblgenomes_file.pl
@@ -277,7 +277,9 @@ foreach my $taxon ( @{ $list_taxon} ) {
 				# change to correct directory
 				chdir $path;
 
-				system("rm $file");
+				if ( -f "$file" ) {
+				      system("rm $file");
+        }
 				# retrieve the file
 				system("wget -t 0 -c -N ftp://$host$ftpdir/$group/$file");
 				#Check size file against DB
